# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Department Projection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projections/{projectionId}/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List overheads
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Add subscription
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /projections/{projectionId}/subscriptions/{id}:
    put:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update subscription
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Remove subscription
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projections/{projectionId}/sales-support:
    get:
      operationId: listSalesSupportResources
      tags: [salesSupport]
      summary: List sales support resources
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesSupportResource"
    post:
      operationId: createSalesSupportResource
      tags: [salesSupport]
      summary: Add sales support resource
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesSupportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSupportResource"
  /projections/{projectionId}/sales-support/{id}:
    put:
      operationId: updateSalesSupportResource
      tags: [salesSupport]
      summary: Update sales support resource
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesSupportBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSupportResource"
    delete:
      operationId: deleteSalesSupportResource
      tags: [salesSupport]
      summary: Remove sales support resource
      parameters:
        - name: projectionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List all quotations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quotation"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Create a quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationWithLineItems"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get quotation with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationWithLineItems"
    put:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update quotation header
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuotationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationWithLineItems"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotations/{quotationId}/line-items:
    post:
      operationId: createQuotationLineItem
      tags: [quotations]
      summary: Add line item
      parameters:
        - name: quotationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLineItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationLineItem"
  /quotations/{quotationId}/line-items/{id}:
    put:
      operationId: updateQuotationLineItem
      tags: [quotations]
      summary: Update line item
      parameters:
        - name: quotationId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLineItemBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationLineItem"
    delete:
      operationId: deleteQuotationLineItem
      tags: [quotations]
      summary: Remove line item
      parameters:
        - name: quotationId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        projectionId:
          type: integer
        name:
          type: string
        currency:
          type: string
        originalPrice:
          type: number
        sarEquivalent:
          type: number
        yearlySar:
          type: number
        monthlySar:
          type: number
        createdAt:
          type: string
      required: [id, projectionId, name, currency, originalPrice, sarEquivalent, yearlySar, monthlySar, createdAt]
    CreateSubscriptionBody:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        originalPrice:
          type: number
      required: [name, currency, originalPrice]
    UpdateSubscriptionBody:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
        originalPrice:
          type: number
    SalesSupportResource:
      type: object
      properties:
        id:
          type: integer
        projectionId:
          type: integer
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        months:
          type: integer
        marginPercent:
          type: number
        ctc:
          type: number
        totalSalaryCost:
          type: number
        createdAt:
          type: string
      required: [id, projectionId, title, country, salarySar, months, marginPercent, ctc, totalSalaryCost, createdAt]
    CreateSalesSupportBody:
      type: object
      properties:
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        months:
          type: integer
        marginPercent:
          type: number
      required: [title, country, salarySar]
    UpdateSalesSupportBody:
      type: object
      properties:
        title:
          type: string
        country:
          type: string
        salarySar:
          type: number
        months:
          type: integer
        marginPercent:
          type: number
    Quotation:
      type: object
      properties:
        id:
          type: integer
        projectionId:
          type: integer
          nullable: true
        quotationNumber:
          type: string
        companyName:
          type: string
        clientName:
          type: string
        date:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, quotationNumber, companyName, clientName, date, status, createdAt, updatedAt]
    QuotationLineItem:
      type: object
      properties:
        id:
          type: integer
        quotationId:
          type: integer
        sortOrder:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        priceMonthly:
          type: number
        totalMonths:
          type: integer
      required: [id, quotationId, sortOrder, description, quantity, unit, priceMonthly, totalMonths]
    QuotationWithLineItems:
      type: object
      properties:
        id:
          type: integer
        projectionId:
          type: integer
          nullable: true
        quotationNumber:
          type: string
        companyName:
          type: string
        clientName:
          type: string
        date:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuotationLineItem"
      required: [id, quotationNumber, companyName, clientName, date, status, createdAt, updatedAt, lineItems]
    CreateQuotationBody:
      type: object
      properties:
        projectionId:
          type: integer
          nullable: true
        quotationNumber:
          type: string
        companyName:
          type: string
        clientName:
          type: string
        date:
          type: string
        status:
          type: string
      required: [quotationNumber, companyName, clientName, date]
    UpdateQuotationBody:
      type: object
      properties:
        quotationNumber:
          type: string
        companyName:
          type: string
        clientName:
          type: string
        date:
          type: string
        status:
          type: string
    CreateLineItemBody:
      type: object
      properties:
        sortOrder:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        priceMonthly:
          type: number
        totalMonths:
          type: integer
      required: [description]
    UpdateLineItemBody:
      type: object
      properties:
        sortOrder:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        priceMonthly:
          type: number
        totalMonths:
          type: integer
...