# lib/api-zod/src/generated/types/createLineItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLineItemBody {
  sortOrder?: number;
  description: string;
  quantity?: number;
  unit?: string;
  priceMonthly?: number;
  totalMonths?: number;
}