# lib/api-zod/src/generated/types/createQuotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuotationBody {
  projectionId?: number | null;
  quotationNumber: string;
  companyName: string;
  clientName: string;
  date: string;
  status?: string;
}