# lib/api-zod/src/generated/types/quotationWithLineItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */
import type { QuotationLineItem } from "./quotationLineItem";

export interface QuotationWithLineItems {
  id: number;
  projectionId?: number | null;
  quotationNumber: string;
  companyName: string;
  clientName: string;
  date: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  lineItems: QuotationLineItem[];
}