# lib/api-zod/src/generated/types/salesSupportResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Department Projection API
 * OpenAPI spec version: 0.1.0
 */

export interface SalesSupportResource {
  id: number;
  projectionId: number;
  title: string;
  country: string;
  salarySar: number;
  months: number;
  marginPercent: number;
  ctc: number;
  totalSalaryCost: number;
  createdAt: string;
}